requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
psycopg2-binary>=2.9.9
voyageai>=0.2.0
anthropic>=0.20.0
python-dotenv>=1.0.0
slowapi>=0.1.9
requests>=2.31.0
pymupdf>=1.23.0
pydantic>=2.0.0
